+++ dune/typetree/leafnode.hh
@@ -37,6 +37,7 @@ namespace Dune {
       static const bool isComposite = false;
 
       //! Leafs have no children.
+      [[deprecated("Will be removed after release 2.9. Use degree()")]]
       static const std::size_t CHILDREN = 0;
 
       //! The type tag that describes a LeafNode.